(cc-back-block) (cc-safe-uplist) (cc-safe-downlist): obsolete
(cc-search-uplist-for-classkey): completely rewritten for speed. it
now uses re-search-backwards limited to two bod's back.  Orders of
magnitude faster when operating towards the end of a large file.  In
fact exec time is not buffer size dependent anymore (but like
everything else, it is bod dependent of course).

(cc-guess-basic-semantics): fixed cases 4, 7, and 12 to work with the
new cc-search-uplist-for-classkey