(cc-safe): added this macro back in -- it seems to be useful
(cc-beginning-of-statement): really improved this a lot. we only go
back a sexp if we aren't already at the beginning of indentation.
This is now a more generally useful function since it seems to do the
right thing under all circumstances.

(cc-guess-basic-semantics): lots of changes to use the new cc-b-o-s
function.  All known indentations after hanging braces are now fixed.
This includes case 4D.5, CASE 7, CASE 12, and CASE 13.F.  Are there
others?