(c-lineup-comment): renamed from c-indent-for-comment
(c-lineup-comment): call c-comment-indent when a comment-only line's
indentation is at or to the right of comment-column.

(c-guess-basic-semantics): when a comment-only line's indentation is
at or to the right of comment-column, only 'comment-intro is on the
semantics list since the comment does not indent relative to a
non-comment line would get indented.