(c-guess-basic-syntax): installed this patch:
In ObjC mode 4.6 the tests 6D and 6E in c-guess-basic-syntax need to
be swapped. Otherwise what is a continued method call line will be
recognized as a 'arglist-cont-nonempty. Swapping these test will not
interfere with the normal behaviour of cc-mode, since the test for the
method call is surrounded with a test on (eq major-mode 'objc-mode).

Emanuele (lele) Gaifax     | ...calling Emacs convenient, of course, is
lele@nautilus.sublink.org  | like calling oxygen useful   -- Rens Troost