(c-file-style, c-file-offsets): New variables.
(c-find-file-hook): New function.

These new variables and function implement the ability to set both
styles and offsets in a file's Local Variable list.  The obviates the
need to use the functional interface for file local variables, which
is very inconvenient.  This also provides an extremely flexible
interface due to the semantics of setting order (c-file-style first,
then c-file-offsets).

This will only work for Emacs 19's.