(c-crosses-statement-barrier-p): New function to help support rewrite of c-beginning-of-statement-1.
(c-backward-to-start-of-do): Semantics change. If matching do is
found, t is returned.  Otherwise, point is left unchanged and nil is
returned.

(c-beginning-of-statement-1, c-guess-basic-syntax): Declarations in
switch statements before first case label now work properly.  Major
semantics change on the relpos's for statements.  Statements now are
always indented relative to their c-beginning-of-statement (i.e. the
statement just before this line).  This works correctly even in the
face of nested conditionals and substatements.  This is a change for
the better so users can modify a line of code and have subsequent
statements line up under the modification.  Before, statements lined
up under the first statement in a block.