(c-recongize-knr-p, c++-mode, java-mode): Made into a non-user configurable variable.  This variable is always t for C and Objective-C modes and nil for C++ and Java modes.
(c-guess-basic-syntax): CASE 5G: Although we still test on
c-recognize-knr-p, this is no longer user configurable, and no longer
imposes potentially disasterous impact on performance.  Adopt
c-mode.el assumption that argdecls are indented at least one space and
that function headers are not indented.  This prevents us from
potentially searching back to the top of the file when looking at a
large number of top-level delarations.