(c-Java-defun-prompt-regexp): new variable
(java-mode): setq defun-prompt-regexp to c-Java-defun-prompt-regexp,
as per Bob Weiner's <weiner@infodock.com> suggestion:

In cc-mode's Java setup, definitions are not properly handled since
the opening brace of most definitions does not occur in the first column.
Emacs provides a variable, defun-prompt-regexp, to handle such problems
created by ugly programming style.  Sun of course had to make the matching
regexp for Java be inordinately complex, so complain to them about
how the following looks.

As part of the java-mode initialization, the following setting should
be used.  (This is not the original source form of the regexp since it
comes from the OO-Browser, but using that would require definition of another
6 or more variables.)  This expression does work as I've tested it with
my syntactic selection package, id-select.el.  Moving forward and backward
by sexps still won't move you by a whole method, however, since both ()
and {} are used in computing Java sexpressions, and so argument lists get in
the way.