(c-guess-basic-syntax): Changes to fix broken member-init recognition when colon is embedded in the middle of a line.  Also fixes to cascading problems caused by this fix, including a bug that has existed for a while in calculating the relpos of multiline class instance variable declarations.
CASE 5D: Don't go to the beginning of the line.  Instead cruise
backwards by measured sexps.  CASE 5D.1: Placehold the location
discovered before the colon, for use in 2nd half of the and.  In the
2nd half of the and, watch out for c-class-key which could indicate
we're looking at an inheritance specification instead of a member init
spec.  CASE 5D.2: We're left at a different buffer position when we
get to this test, so we need to find the location of the c-inher-key
first before we see if we're looking at it. CASE 5D.4: We need to
placehold the location of the `<' character that starts the template
since later, we're left at a different buffer location than we used to
be.