Improved recovery of syntactic errors:
(c-parsing-error): Changed this variable to hold the message for
any syntactic error that is discovered.

(c-parse-state): Search backward from point instead of the bod
position when the latter is invalid.  This makes CC Mode recover
faster when there are unbalanced close braces.

(c-backward-to-start-of-if): Use c-parsing-error to report dangling
"else" clauses instead of throwing an error, and fall back to a
reasonable position.

(c-indent-line): Added argument to avoid reporting syntactic errors.

(c-show-syntactic-information): Don't report any syntactic errors.