Several fixes to treat macros as code and not literals and to handle line continuations transparently.
(c-in-literal): No longer recognizes macros as literals by default.
An extra flag argument is required to do that.

(c-macro-start, c-query-macro-start, c-query-and-set-macro-start):
Added a cache for the macro start position.

(c-forward-syntactic-ws, c-backward-syntactic-ws): Fixes for more
speed when limits are given.  Workaround for bad behavior in
forward-comment in some emacsen when it hits a buffer limit with a
large repeat count.