(c-at-expression-start-p): New function like `c-at-statement-start-p' that additionally recognizes commas and expression parentheses as delimiters.
(c-at-statement-start-p): The optional COMMA-DELIM argument is gone.
Use `c-at-expression-start-p' instead.

(c-guess-basic-syntax): More consistent anchoring for `arglist-close'.
The main effect is that the anchor position is at the start of the
statement instead of the opening paren even when the opening paren is
at boi.  Thus odd indentation is avoided in that case if
`arglist-close' is set to `+'.