Enhance cc-bytecomp-ignore-obsolete to cope with the missing function `byte-compile-obsolete' in Emacs >= 24.
(top-level): Add variable cc-bytecomp-push-vars.  Set this and similar
variables to nil each time the file is loaded.
(cc-bytecomp-setup-environment, cc-bytecomp-restore-environment): Add
clauses to setup and restore "pushed" variables.
(cc-bytecomp-push): New macro to "push" a variable onto an internal
stack.
(cc-bytecomp-ignore-obsolete): Add a superfluous check to suppress a
compiler warning
(cc-bytecomp-ignore-obsolete): In Emacs >= 24, use
byte-compile-not-obsolete-funcs to mark a function as obsolete.