Handle `parse-partial-sexp' landing inside a comment opener. Also adapt to the new values of element 7 of a parse state.
cc-engine.el (c-state-pp-to-literal): New optional parameter
`not-in-delimiter'.  Handle being inside comment opener.
(c-invalidate-state-cache-1): Reckon with an extra "invalid" character in
case we're typing a '*' after a '/'.
(c-literal-limits): Handle the awkward "not-in-delimiter" cond arm
instead by passing the parameter to c-state-pp-to-literal.

cc-fonts.el (c-font-lock-doc-comments): New handling for elt. 7 of a
parse state.