Fix fontification bugs with constructors and const.
cc-engine.el (c-forward-decl-or-cast-1): (Just after CASE 2) Remove the
check for the absence of a suffix construct after a function declaration
with only types (no identifiers) in the parentheses.  Also, accept a
function declaration with just a type inside the parentheses, if this
type can be positively recognised as such, or if a prefix keyword like
"explicit" nails down the construct as a declaration.